code lint and indentation fixes
This project needs a desperate tune-up. Bear in mind that I haven’t
worked on this project for months and have learned a few new things
since then. I will start working on factorization and module
segmentation in the coming weeks to make this project more friendly for
those who want to contribute. Unit testing and proper documentation is
also needed. Any tips and tricks are gladly appreciated.